app/layout.tsxの中身もFSDに揃えて分離&移動
