+++ app/layout.tsx
@@ -1,23 +1,7 @@
-import { Metadata } from 'next'
-import CommonPageTemplate from '../src/components/CommonPageTemplate'
-import '../src/app/styles/globals.css'
-import ReduxProvider from './_providers/ReduxProvider'
+import type { Metadata } from 'next';
+import { rootMetadata } from '../src/shared/metadata';
+import RootLayout from '../src/shared/ui/layout/RootLayout';
 
-type RootLayoutProps = React.PropsWithChildren
+export const metadata: Metadata = rootMetadata;
 
-export const metadata: Metadata = {
-  title: 'Next.js Template(TodoList)',
-  description: 'This is Next.js Sample Template',
-}
-
-const RootLayout: React.FC<RootLayoutProps> = ({ children }: RootLayoutProps) => (
-  <html lang="ja">
-    <body>
-      <ReduxProvider>
-        <CommonPageTemplate>{children}</CommonPageTemplate>
-      </ReduxProvider>
-    </body>
-  </html>
-)
-
-export default RootLayout
+export default RootLayout;
