+++ src/shared/metadata/index.ts
@@ -1,0 +1,1 @@
+export { metadata as rootMetadata } from './rootMetadata';
