+++ src/shared/metadata/rootMetadata.ts
@@ -1,0 +1,6 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: 'Next.js Template(TodoList)',
+  description: 'This is Next.js Sample Template',
+};
