+++ src/shared/ui/layout/RootLayout.tsx
@@ -1,0 +1,17 @@
+import CommonPageTemplate from '../../../components/CommonPageTemplate';
+import '../../../app/styles/globals.css';
+import ReduxProvider from '../../../../app/_providers/ReduxProvider';
+
+type RootLayoutProps = React.PropsWithChildren;
+
+const RootLayout: React.FC<RootLayoutProps> = ({ children }: RootLayoutProps) => (
+  <html lang="ja">
+    <body>
+      <ReduxProvider>
+        <CommonPageTemplate>{children}</CommonPageTemplate>
+      </ReduxProvider>
+    </body>
+  </html>
+)
+
+export default RootLayout;
